refactor(frontend) componentize requests tables
